requirements.txt
torch>=1.0.0
torchvision~=0.7.0
dominate>=2.3.1
dill~=0.3.3
scikit-image~=0.18.1
scipy~=1.5.2
opencv-python~=4.4.0.46
numpy~=1.18.5
pillow~=7.2.0
pyqt5~=5.15.2
imageio~=2.9.0
flask~=1.1.2